requirements.txt
asgiref==3.2.7
cffi==1.14.0
cryptography==2.9.2
Django==3.0.6
django-rest-knox==4.1.0
djangorestframework==3.11.0
psycopg2==2.8.5
pycparser==2.20
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
whitenoise==5.0.1
